fix: handle last review comments
